Added copyfile method to FileUtilities class for convenience when testing.

+++ src/main/com/google/code/tvrenamer/controller/util/FileUtilities.java
@@ -33,12 +33,40 @@ public class FileUtilities {
 	 * @param observer
 	 *            The observer to notify (can be null).
 	 * @param canStop
-	 *            if true, the copy can be stopped by interrupting the thread
+	 *            if true, the move can be stopped by interrupting the thread
 	 * @return true on success, false otherwise.
 	 * 
 	 * @since jEdit 4.3pre9
 	 */
 	public static boolean moveFile(File source, File dest, ProgressObserver observer, boolean canStop) {
+		return doAction(source, dest, observer, canStop, true);
+	} // }}}
+
+	// {{{ copyFile() method, based on the moveFile() method above
+	/**
+	 * Copies the source file to the destination.
+	 * 
+	 * If the destination cannot be created or is a read-only file, the method returns <code>false</code>. Otherwise,
+	 * the contents of the source are copied to the destination and <code>true</code> is returned.
+	 * 
+	 * @param source
+	 *            The source file to copy.
+	 * @param dest
+	 *            The destination where to copy the file.
+	 * @param observer
+	 *            The observer to notify (can be null).
+	 * @param canStop
+	 *            if true, the copy can be stopped by interrupting the thread
+	 * @return true on success, false otherwise.
+	 * 
+	 * @since jEdit 4.3pre9
+	 */
+	public static boolean copyFile(File source, File dest, ProgressObserver observer, boolean canStop) {
+		return doAction(source, dest, observer, canStop, false);
+	} // }}}
+
+	private static boolean doAction(File source, File dest, ProgressObserver observer, boolean canStop,
+		boolean deleteOnSuccess) {
 		boolean ok = false;
 
 		if ((dest.exists() && dest.canWrite()) || (!dest.exists() && dest.getParentFile().canWrite())) {
@@ -55,9 +83,9 @@ public class FileUtilities {
 				IOUtilities.closeQuietly(fis);
 			}
 
-			if (ok)
+			if (ok && deleteOnSuccess)
 				source.delete();
 		}
 		return ok;
-	} // }}}
+	}
 }
